==> scripts/no_strat/4_analisis_exploratorio-geom.Rmd <==
---
title: "Análisis Exploratorio"
author: "Juan Baeza Ruiz-Henestrosa"
date: "`r Sys.Date()`"
output: pdf_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(skimr)
library(sf)
library(corrplot)
library(GGally) # Coordenadas paralelas
library(ggpubr)
```


Cargamos el archivos de datos depurados:

```{r}
# load("salidas_intermedias/datos_depurados_geom_2024-04-23.RData")
load("salidas_intermedias/datos_depurados_geom_2024-04-26.RData")
load("salidas_intermedias/datos_strat_depurados_geom_2024-04-27.RData")


```


## Resumen numérico

```{r}
datos %>%  st_drop_geometry %>% skim()
```

Se tienen variables numéricas y categóricas. Las variables numéricas presentan cada una distinta variabilidad, lo que sugiere que probablemente sea conveniente trabajar con los datos tipificados.



## Variable objetivo:

### Distribución temporal
#### Mes
```{r}
datos %>% 
  st_drop_geometry() %>% 
  # filter(fire==1) %>% 
  count(month(date,label=T),fire) %>% 
  rename("mes" = "month(date, label = T)") %>% 
  ggplot(aes(x = mes,y = n,fill=fire)) +
  geom_col(position="dodge",alpha=0.8) 
```

Se observa que los casos positivos se concentran entre junio y septiembre (estación de incendios) mientras que los casos negativos están uniformemente distribuidos entre los distintos meses del año.



#### Año
```{r}
datos %>% 
  st_drop_geometry() %>% 
  # filter(fire==1) %>% 
  count(year(date),fire) %>% 
  rename("año" = "year(date)") %>% 
  ggplot(aes(x = año,y = n,fill=fire)) +
  geom_col(position="dodge",alpha=0.8) +
  scale_x_continuous(breaks=2002:2022) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) +
  geom_smooth(se=F)

# En 2007 no se han generado observaciones positivas
# Causa: En el shapefile que contiene los incendios producidos en 2007 solo hay 4 observaciones y ninguna tenía la fecha, por lo que estas no han podido utilizarse

# Valores muy bajos en 2008 y 2010, revisar
# 2010: correcto, en el archivo con los polígonos de incendios en 2010 solo hay 26 observaciones
# 2008: en el archivo relativo a ese año hay 37 observaciones, pero 11 de ellas no tienen la fecha de inicio, por lo que no se han podido utilizar
```


#### Día
```{r}
datos %>% 
  st_drop_geometry() %>% 
  # filter(fire==1) %>% 
  count(weekdays(date),fire) %>% 
  rename("dia" = "weekdays(date)") %>% 
  ggplot(aes(x = dia,y = n,fill=fire)) +
  geom_col(position="dodge",alpha=0.8) +
  scale_x_discrete(limits=c("lunes", "martes", "miércoles", "jueves", "viernes", "sábado", "domingo")) +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=0.5)) +
  geom_smooth(se=F)
```
Podría dar la impresión de que más incendios ocurren en fin de semana

### Distribución espacial
```{r}
library(mapSpain)

and <- esp_get_ccaa(ccaa = "Andalucía") %>% st_transform(st_crs(datos))
prov <- esp_get_prov() %>% filter(nuts2.name=="Andalucía") %>% st_transform(st_crs(datos))

# Casos positivos
g1 = ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos %>% filter(fire==1),size = 1, alpha = 0.4,col="red")+
  ggtitle("Distribución espacial de casos positivos")

# Casos negativos
g0 = ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos %>% filter(fire==0),size = 1, alpha = 0.4,col="blue") +
  ggtitle("Distribución espacial de casos negativos")

ggarrange(g1,g0,nrow=2)
```


## Análisis univariantes

### T2M
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=T2M),size = 1.2, alpha = 0.6) +
  scale_color_gradientn(colours = rainbow(5,rev=T)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de T2M")
```

```{r}
datos %>% 
  st_drop_geometry() %>% 
  group_by(month(date,label=T),fire) %>% 
  summarize(AVG_T2M = mean(T2M),
            Mes = `month(date, label = T)`) %>% 
  ggplot(aes(x=Mes,y=AVG_T2M,fill=fire)) +
  geom_col(position="dodge",alpha=0.8)
```

### RH2M
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=RH2M),size = 1.2, alpha = 0.6) +
  scale_color_gradientn(colours = rainbow(5,rev=F)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de RH2M")
```

```{r}
datos %>% 
  st_drop_geometry() %>% 
  group_by(month(date,label=T),fire) %>% 
  summarize(AVG_RH2M = mean(RH2M),
            Mes = `month(date, label = T)`) %>% 
  ggplot(aes(x=Mes,y=AVG_RH2M,fill=fire)) +
  geom_col(position="dodge",alpha=0.8)
```

### GWETTOP

```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=GWETTOP),size = 1.2, alpha = 0.6) +
  scale_color_gradientn(colours = rainbow(5,rev=F)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de GWETTOP")
```

```{r}
datos %>% 
  st_drop_geometry() %>% 
  group_by(month(date,label=T),fire) %>% 
  summarize(AVG_GWETTOP = mean(GWETTOP),
            Mes = `month(date, label = T)`) %>% 
  ggplot(aes(x=Mes,y=AVG_GWETTOP,fill=fire)) +
  geom_col(position="dodge",alpha=0.8)
```

### WS10M
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=WS10M),size = 1.2, alpha = 0.6) +
  scale_color_gradientn(colours = rainbow(5,rev=T)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de WS10M")
```

```{r}
datos %>% 
  st_drop_geometry() %>% 
  group_by(month(date,label=T),fire) %>% 
  summarize(AVG_WS10M = mean(WS10M),
            Mes = `month(date, label = T)`) %>% 
  ggplot(aes(x=Mes,y=AVG_WS10M,fill=fire)) +
  geom_col(position="dodge",alpha=0.8)
```

### WD10M
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=WD10M),size = 1.2, alpha = 0.6) +
  # scale_color_gradientn(colours = rainbow(5,rev=T)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de WD10M")
```

```{r}
# Quisiera tener la dirección mayoritaria por mes y si hay incendio o no 

mode <- function(x) { names(which.max(table(x))) }

datos_mode_WD10M = datos %>%
  st_drop_geometry() %>%
  group_by(month(date,label=T),fire) %>%
  summarize(mode_WD10M = mode(WD10M)) 

tibble(mes = datos_mode_WD10M[datos_mode_WD10M$fire==1,1],
       fire1 = datos_mode_WD10M[datos_mode_WD10M$fire==1,3],
       fire0 = datos_mode_WD10M[datos_mode_WD10M$fire==0,3]) 

# En realidad de esto creo que no saco nada
```


### PRECTOTCORR

```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=PRECTOTCORR),size = 1.2, alpha = 1) +
  scale_color_gradientn(colours = rainbow(5,rev=T)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de PRECTOTCORR")
```

```{r}
datos %>% 
  st_drop_geometry() %>% 
  group_by(month(date,label=T),fire) %>% 
  summarize(AVG_PRECTOTCORR = mean(PRECTOTCORR),
            Mes = `month(date, label = T)`) %>% 
  ggplot(aes(x=Mes,y=AVG_PRECTOTCORR,fill=fire)) +
  geom_col(position="dodge",alpha=0.8)
```


 <!-- IDEA PARA PODER EVALUAR MEJOR EL RENDIMIENTO REAL DEL MODELO: ESTRATIFICAR POR MESES -->
Ya que el objetivo no es que el modelo sea capaz de clasificar muy bien los datos del dataset, si no su utilidad práctica de cara predecir incendios forestales.
EL objetivo es que la muestra permita detectar los patrones que causan los incendios forestales, por lo que el objetivo al generar las muestras negativas es evitar en la mayor medida posible los sesgos debidos al proceso de selección de la muestra. Se busca que el modelo sea capaz de discriminar las observaciones en riesgo de incendio forestal, no los sesgos introducidos en el proceso de generación de la muestra.

<!-- Al generar los ejemplos negativos, mi objetivo es intentar evitar los sesgos que se puedan producir en el proceso de selecciom, con el objetivo de que la diferencia entre ejemplos positivos y negativos sea realmente debida a las carácterísticas de los incendios y no a sesgos introducidos en el proceso de muestreo. -->

### NDVI
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=NDVI),size = 1.2, alpha = 0.6) +
  scale_color_gradientn(colours = rainbow(5,rev=T)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de NDVI")
```

```{r}
datos %>% 
  st_drop_geometry() %>% 
  group_by(month(date,label=T),fire) %>% 
  summarize(AVG_NDVI = mean(NDVI),
            Mes = `month(date, label = T)`) %>% 
  ggplot(aes(x=Mes,y=AVG_NDVI,fill=fire)) +
  geom_col(position="dodge",alpha=0.8)
```


### poblacion

```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=poblacion),size = 1.2, alpha = 0.6) +
  scale_color_gradientn(colours = rainbow(5,rev=T)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de poblacion")
```

```{r}
# Este en realidad no sirve para nada, pero es curioso

datos %>% 
  st_drop_geometry() %>% 
  group_by(month(date,label=T),fire) %>% 
  summarize(AVG_poblacion = mean(poblacion),
            Mes = `month(date, label = T)`) %>% 
  ggplot(aes(x=Mes,y=AVG_poblacion,fill=fire)) +
  geom_col(position="dodge",alpha=0.8)
```


### dens_poblacion
```{r}
ggplot(data = prov) + 
  geom_sf() +
  geom_sf(data = datos, aes(color=dens_poblacion),size = 1.2, alpha = 0.6) +
  scale_color_gradientn(colours = rainbow(5,rev=T)) +
  # scale_color_gradient(low="blue", high="red")+
  ggtitle("Distribución espacial de dens_poblacion")
```





## Variables numéricas


18 variables numéricas

```{r}
datos_numeric = datos |> select(where(is.numeric)) %>% st_drop_geometry()

# datos_numeric$fire = as.numeric(datos$fire)
```


### Correlaciones:

```{r}
R = cor(datos_numeric)

corrplot(R,method = "ellipse",type = "upper")

summary(R-diag(diag(R)))
```

Las variables más correlacionadas son T2M con RH2M (negativamente, -0.79), T2M con GWETTOP (negativamente, -0.77) y GWETTOP con R2HM (positivamente, 0.78). Tiene sentido que la humedad del aire a dos esté correlacionada positivamente con la humedad del suelo y que ambas estén -egativamente correlacionadas con la temperatura del aire.


<!-- Esto puede no tener sentido siendo estrictos -->
```{r}
cor(datos_numeric,as.numeric(datos$fire))
```
Las variables más correladas con la variable respuesta son T2M, GWETTOP, RH2M.


### Gráfico de coordenadas paralelas
```{r}
datos |> 
  select(fire,where(is.numeric)) |> 
  ggparcoord(columns=2:19,alphaLines=0.1,groupColumn = "fire") +
  xlab('') + ylab('') + scale_x_discrete(labels=colnames(datos_numeric)) +
  ggtitle("Tipificación a normal estándar (por defecto)") +
  theme(plot.title = element_text(size = 9),
        axis.text.x = element_text(angle = 90))
```


Considerando las variables tipificadas, se observa que aquellas con más variabilidad son PRECTOTCORR, población, dist_poblacion y dist_camino. Todas estas variables tienen un sesgo positivo, lo cual es razonable pues todas ellas son variables que solo toman valores positivos.

Se puede que la variable PRECTOTCORR presenta valores atípicamente altos, que se dan en observaciones en las que no se ha observado incendio. 

```{r}
datos |> 
  ggplot(aes(x=fire,y=PRECTOTCORR)) +
  geom_boxplot()
```

```{r}
by(datos$PRECTOTCORR,datos$fire,summary)
```

Acumular 59.90mm de lluvia en un día es un valor posible, probablemente atípico en esta región pero no hay motivo para considerar que es un valor erróneo. Es coherente que las observaciones en las que se han registrado mayores niveles de precipitación sean observaciones negativas.

En el gráfico también observamos que hay valores atípicamente altos de la variable población.
Se puede observar que los valores más altos de elevacion, de dist_carretera y de dist_electr se corresponden con observaciones negativas.

```{r}
g0 = datos |> 
  ggplot(aes(x=fire)) 

g1 = g0 +
  geom_boxplot(aes(y=elevacion))

g2 = g0 +
  geom_boxplot(aes(y=dist_carretera))


g3 = g0 +
  geom_boxplot(aes(y=dist_electr))


ggarrange(g1,g2,g3)

```

### Matriz de dispersion
```{r}
datos_numeric |>  
  mutate(fire = datos$fire) |> 
  ggpairs(aes(color=fire))
```


### Box-plot

Se hace primero el boxplot con todas las variables:

```{r}
boxplot(datos_numeric)
```


```{r}
boxplots <- map(names(datos_numeric), ~ ggplot(datos, aes(x = fire, y = .data[[.x]])) +
                geom_boxplot() +
                labs(title = paste(.x, "~ fire")))

ggarrange(plotlist = boxplots,ncol=4,nrow=5)
# do.call(ggarrange,c(boxplots,ncol = 3,nrow=6))
```


## Variables categóricas

Las variables date, cod_municipio y municipio son "identificadoras". Las variables explicativas categóricas son WD10M, orientacion, enp y uso_suelo. 


```{r}
datos_categ <- datos |> select(WD10M,orientacion,enp,uso_suelo) %>% st_drop_geometry()
```

### Histogramas

```{r}
histogramas <- map(names(datos_categ), ~ ggplot(datos, aes(x = .data[[.x]], fill = fire)) +
                    geom_bar(position = "dodge", alpha = 0.8) +
                    labs(title = .x)+theme(axis.text.x = element_text(size=7)))

ggarrange(plotlist = histogramas, ncol = 2, nrow = ceiling(length(histogramas)/2))

```

WD10M: viento de sur - oeste - norte -> más observaciones positivas

orientación: orientacion este - sur - suroeste -> Más observaciones positivas

enp: no hay diferencias observables

uso_suelo: no se pueden sacar conclusiones pues las diferencias se deben al procedimiento de obtención de los datos




## PCA

### Solo variables numericas
```{r}
pca <- princomp(datos_numeric,cor=T)

summary(pca)
```

Se necesitan 11 componentes principales para explicar al menos el 80% de la variabilidad de los datos (tipificados).

### Todas las variables
```{r}
# Usamos model.matrix para trnasformar los factores usando variables dummies:
M = model.matrix(fire~.-1,data=datos |> st_drop_geometry() |> select(-c(date,municipio,cod_municipio)))

pca <- princomp(M,cor=T)

summary(pca)
```

Se necesitan 30 componentes principales para explicar el 80% de la varianza de los datos.

## Cluster

### Variables numéricas
Se aplica un algoritmo de clustering basado en k-medias adaptado para grandes conjuntos de datos.

```{r}
library(cluster)
clara<- clara(datos_numeric,2)
table(clara$clustering)

# Si los datos no están normalizados, el método no es capaz de separar los datos
```

Vemos que la partición obtenida con los datos a penas consigue separarlos. Se debe a que algunas variables tienen mucha más variabilidad que otras y k-medias no funciona bien cuando las variables tienen varianzas parecidas. 

Probamos a normalizarlos antes de aplicar el algoritmo.

```{r}
set.seed(12345)
# Probamos tipificando los datos:
clara = datos_numeric |> 
  scale() |> 
  clara(k=2)

table(clara$clustering)
# Ahora el algoritmo consigue separar mejor los datos

t = table(clara$clustering,datos$fire)
t = rbind(t[2,],t[1,])
rownames(t) <- paste0("cluster_",c(1,2))
colnames(t) <- c("No fire","Fire")
t
# addmargins(t)

prop.table(t,2)*100

sum(diag(t))/nrow(datos) # Proporción de observaciones "correctamente separada"
```
Vemos que obtenemos la partición generada con los datos normalizado si está aportando información con respecto a la variable respuesta. El 84% de las observaciones en las que ha habido un incendio se encuentran en el cluster 2, mientras que el 50% de las observaciones en las que no ha habido incendios se encuentran en el cluster 1.


### Todas las variables

```{r}
# Probamos tipificando los datos:
clara = M |> 
  scale() |> 
  clara(k=2)

table(clara$clustering)
# Ahora el algoritmo consigue separar mejor los datos

t = table(clara$clustering,datos$fire)
rownames(t) <- paste0("cluster_",c(1,2))
colnames(t) <- c("No fire","Fire")
t
# addmargins(t)

prop.table(t,2)*100

sum(diag(t))/nrow(datos) # Proporción de observaciones "correctamente separada"
```

No parece aportar nada en el análisis.
